Add facade handlers for app routes

+++ packages/mobile-app/data/wallet/db.ts
@@ -1,10 +1,15 @@
+import { AccountImport } from "@ironfish/sdk/build/src/wallet/walletdb/accountValue";
 import { Kysely, Generated, Migrator } from "kysely";
-import { ExpoDialect, ExpoMigrationProvider, SQLiteType } from "kysely-expo";
+import { ExpoDialect, ExpoMigrationProvider, SQLiteType} from "kysely-expo";
+import * as SecureStore from 'expo-secure-store';
+import { AccountFormat, decodeAccount, encodeAccount } from "@ironfish/sdk";
 
 interface AccountsTable {
     id: Generated<number>;
     name: string;
+    publicAddress: string;
     viewOnlyAccount: string;
+    viewOnly: boolean;
 }
   
 interface Database {
@@ -37,8 +42,10 @@ export class WalletDb {
                       .addColumn("id", "integer", (col) =>
                         col.primaryKey().autoIncrement()
                       )
-                      .addColumn("name", SQLiteType.String, (col) => col.notNull())
+                      .addColumn("name", SQLiteType.String, (col) => col.notNull().unique())
+                      .addColumn("publicAddress", SQLiteType.String, (col) => col.notNull().unique())
                       .addColumn("viewOnlyAccount", SQLiteType.String, (col) => col.notNull())
+                      .addColumn("viewOnly", SQLiteType.Boolean, (col) => col.notNull())
                       .execute();
                   },
                 },
@@ -51,16 +58,28 @@ export class WalletDb {
         return new WalletDb(db);
     }
 
-    async createAccount(name: string, viewOnlyAccount: string) {
-      const result = await this.db.insertInto("accounts").values({
-        name: name,
+    async createAccount(account: AccountImport) {
+      const viewOnlyAccount = encodeAccount({...account, spendingKey: null}, AccountFormat.Base64Json);
+
+      const accountValues = {
+        name: account.name,
+        publicAddress: account.publicAddress,
         viewOnlyAccount: viewOnlyAccount,
-      }).executeTakeFirst();
+        viewOnly: account.spendingKey !== null,
+      }
+
+      const result = await this.db.insertInto("accounts").values(accountValues).executeTakeFirst();
+
+      if (account.spendingKey) {
+        await SecureStore.setItemAsync(account.publicAddress, account.spendingKey, {
+          keychainAccessible: SecureStore.WHEN_UNLOCKED_THIS_DEVICE_ONLY,
+          requireAuthentication: false,
+        });
+      }
 
       return {
         id: Number(result.insertId),
-        name: name,
-        viewOnlyAccount: viewOnlyAccount,
+        ...accountValues,
       };
     }
 
@@ -72,11 +91,36 @@ export class WalletDb {
       return await this.db.selectFrom("accounts").selectAll().where('accounts.name', '==', name).executeTakeFirst();
     }
 
+    async getSpendingKey(publicAddress: string) {
+      return await SecureStore.getItemAsync(publicAddress, {
+        keychainAccessible: SecureStore.WHEN_UNLOCKED_THIS_DEVICE_ONLY,
+        requireAuthentication: false,
+      })
+    }
+
     async renameAccount(name: string, newName: string) {
       return await this.db.updateTable("accounts").where('accounts.name', '==', name).set('accounts.name', newName).executeTakeFirst();
     }
 
     async removeAccount(name: string) {
-      return await this.db.deleteFrom("accounts").where('accounts.name', '==', name).executeTakeFirst();
+      const account = await this.getAccount(name)
+      if (!account) {
+        throw new Error(`No account found with name ${name}`)
+      }
+
+      const result = await this.db.deleteFrom("accounts").where('accounts.name', '==', name).executeTakeFirst();
+
+      if (result.numDeletedRows > 0) {
+        try {
+          await SecureStore.deleteItemAsync(account.publicAddress, {
+            keychainAccessible: SecureStore.WHEN_UNLOCKED_THIS_DEVICE_ONLY,
+            requireAuthentication: false,
+          })
+        } catch {
+          console.log(`Failed to delete spending key for account ${name}`)
+        }
+      }
+
+      return result;
     }
 }
